Move TeamSeasonCollection data structure to SeasonCollection
