--- lib/team_season.rb
@@ -1,24 +1,0 @@
-class TeamSeason
-  attr_reader :game_id,
-              :season,
-              :type,
-              :date_time,
-              :away_team_id,
-              :home_team_id,
-              :away_goals,
-              :home_goals,
-              :venue,
-              :venue_link
-
-  def initialize(row)
-    @game_id = row[:game_id]
-    @season = row[:season]
-    @type = row[:type]
-    @date_time = row[:date_time]
-    @home_team_id = row[:home_team_id]
-    @away_team_id = row[:away_team_id]
-    @away_goals = row[:away_goals]
-    @home_goals = row[:home_goals]
-    @venue = row[:venue]
-  end
-end
