+++ lib/tracker.rb
@@ -8,13 +8,10 @@ require_relative 'game_collection'
 require_relative 'team_collection'
 require_relative 'game_teams_collection'
 require_relative 'season_collection'
-require_relative 'team_season'
-require_relative 'team_season_collection'
 
 class Tracker
   attr_reader :game_collection,
               :team_collection,
-              :team_season_collection,
               :season_collection,
               :game_teams_collection
 
@@ -30,7 +27,6 @@ class Tracker
     @game_collection = GameCollection.new(games)
     @team_collection = TeamCollection.new(teams)
     @season_collection = SeasonCollection.new(games)
-    @team_season_collection = TeamSeasonCollection.new(games)
     @game_teams_collection = GameTeamsCollection.new(game_teams)
   end
 end
